Implement back publishing support
**Problem**
For plugin code, like sbt plugins and Scala compiler plugin,
and even normal libraries, it's common for the maintainers to
want to back publish the current code base against a new version
of Scala (or sbt, Scala Nave etc) instead of bumping the version.

The current "bump the version" approach effectively pushes the work
of figuring out the correct patch version number to the end users.
In addition, if the plugin or the library does not cross publish
all variants, then the end user might need to figure out different
patch version to use effectively the same code.

**Solution**
This implements a mini DSL for tag `version[@command|@3.3.4][#comment]`,
which allows plugin authors to back publish a code base.